+++ .travis.yml
@@ -2,6 +2,9 @@ sudo: required
 os:
   - linux
 
+services:
+  - docker
+
 language: node_js
 node_js:
   - "6"
