+++ appveyor.yml
@@ -4,6 +4,9 @@ environment:
     - nodejs_version: "6"
     - nodejs_version: "7"
 
+image:
+  - Visual Studio 2017
+
 platform:
   - x86
   - x64
